Support pattern matching with sets (#391)
* support set pattern
  * let {} = {}; 1
  * let {a, 42} = {42, 1}; a
  * let {1, 2, 3, ...t} = {1, 2, 3, 42, 43}; t
  * let x = 1; let y = 42; let {(x), (y), ...t} = {1, 42, 5, 6}; t

* Add ExprPattern

* restructure pattern code
