+++ syntax/expr_let_test.go
@@ -84,6 +84,26 @@ func TestExprLetDictPattern(t *testing.T) {
 	AssertCodePanics(t, `let {"x": a, "x": a} = {"x": 4, "x": 4}; a`)
 }
 
+func TestExprLetSetPattern(t *testing.T) {
+	t.Parallel()
+	AssertCodesEvalToSameValue(t, `1`, `let {} = {}; 1`)
+	AssertCodesEvalToSameValue(t, `1`, `let {42} = {42}; 1`)
+	AssertCodesEvalToSameValue(t, `1`, `let {a} = {1}; a`)
+	AssertCodesEvalToSameValue(t, `1`, `let {a, 42} = {42, 1}; a`)
+	AssertCodesEvalToSameValue(t, `{1, 42}`, `let {...t} = {1, 42}; t`)
+	AssertCodesEvalToSameValue(t, `{42, 43}`, `let {1, 2, 3, ...t} = {1, 2, 3, 42, 43}; t`)
+	AssertCodesEvalToSameValue(t, `5`, `let x = 1; let y = 42; let {(x), (y)} = {42, 1}; 5`)
+	AssertCodesEvalToSameValue(t, `{5, 6}`, `let x = 1; let y = 42; let {(x), (y), ...t} = {1, 42, 5, 6}; t`)
+
+	AssertCodeErrors(t, `let {} = {1}; 1`, "")
+	AssertCodeErrors(t, `let {1} = {}; 1`, "")
+	AssertCodeErrors(t, `let {42} = {2}; 1`, "")
+	AssertCodeErrors(t, `let {42, 43}={41, 42}; 1`, "")
+	AssertCodeErrors(t, `let {x, y}={41, 42}; 1`, "")
+	AssertCodeErrors(t, `let {x, ...t}={41, 42}; 1`, "")
+	AssertCodeErrors(t, `let x = 1; let y = 42; let {(x), (y)} = {1, 4}; 2`, "")
+}
+
 func TestExprLetExtraElementsInPattern(t *testing.T) {
 	t.Parallel()
 	AssertCodesEvalToSameValue(t, `42`, `let [...] = [1, 2]; 42`)
